refactor(namespaces): correct misplaced types in library namespaces
Audit of all library types (src/, excluding tests/examples) surfaced several
declarations whose namespace did not match the type's role, shipping assembly, or
folder. Relocated the confirmed cases and cascaded the reference/using updates
through tests, examples and docs. Build is green on every target framework
(netstandard2.0, net48, net6.0, net8.0, net10.0) across the three libraries, the
test project, and all buildable example projects.

Moves (public API surface, source-breaking for consumers):
- NotifiyingJsonSettings: Nucs.JsonSettings.Examples -> Nucs.JsonSettings.
  A shipped, public base class that the Autosave and NotifyChanges packages depend
  on was sitting in an "Examples" namespace; it is production API, not a sample,
  and was the only library type in that namespace.
- FluentJsonSettings: Nucs.JsonSettings.Fluent -> Nucs.JsonSettings.
  The primary fluent entry-point surface now sits in the root namespace.
- ReflectionHelpers: Microsoft.Xna.Framework.Utilities -> Nucs.JsonSettings.
  Vendored internal helper that had kept its third-party origin namespace inside
  the assembly. Remains internal.

Deliberately left in place after review:
- KeySize and EncryptionAlgorithm stay in Nucs.JsonSettings (root). They are the
  arguments to the root-level WithEncryption / WithEncryptionRawKey fluent overloads
  (EncryptionAlgorithm algorithm, KeySize keySize = KeySize.Aes256), so root keeps a
  single `using Nucs.JsonSettings;` sufficient for the entire encrypted-load call and
  honours the shipped release-note compatibility promise. Co-locating them with
  EncryptionModule under .Modulation was considered and rejected on those grounds.
- ISavable / IEncryptedSavable stay in Nucs.JsonSettings (core contracts; moving them
  under the optional .Autosave feature namespace was rejected).
- SecureStringExt stays public in System.Security (intentional extension surface,
  consumed by an example) and ReflectionHelper stays public in
  Nucs.JsonSettings.Reflection (consumed by the NotifyChanges assembly, which is not
  an InternalsVisibleTo friend, so it cannot be made internal).

Cascade:
- Library internals: removed now-dead usings (Nucs.JsonSettings.Examples,
  Nucs.JsonSettings.Fluent, Microsoft.Xna.Framework.Utilities) from JsonSettings.cs,
  Activation.cs and the Autosave/NotifyChanges runtimes; moved types resolve via the
  parent namespace.
- Tests: dropped the dead usings from 37 files.
- Examples: dropped the dead Nucs.JsonSettings.Fluent usings (methods resolve via the
  parent namespace).

Docs reconciled to the new namespaces: README, the website docs (notifications,
encryption, versioning, recovery), api/index.md, the Nucs.JsonSettings overwrite,
and filterConfig.yml. The DocFX-generated api/*.yml and toc.yml are produced by
`docfx metadata` from the assembly and require a regeneration to reflect the moves;
they are not hand-edited.
