+++ examples/JsonSettings.Examples.UI.Avalonia/Settings/DemoSettings.cs
@@ -1,6 +1,5 @@
 using System.Collections.ObjectModel;
 using Nucs.JsonSettings.Autosave;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation.Recovery;
 using Nucs.JsonSettings.NotifyChanges;
