+++ examples/JsonSettings.Examples.UI.WinForms/Settings/VaultSettings.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel;
 using Nucs.JsonSettings.Autosave;
-using Nucs.JsonSettings.Fluent;
 
 namespace Nucs.JsonSettings.Examples.UI.WinForms;
