+++ tests/JsonSettings.Tests/Autosave/AutosaveNotificationsTests.cs
@@ -5,7 +5,6 @@ using System.Runtime.CompilerServices;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nucs.JsonSettings.Autosave;
-using Nucs.JsonSettings.Examples;
 using Nucs.JsonSettings.Tests.Utils;
 
 namespace Nucs.JsonSettings.Tests.Autosave {
