+++ tests/JsonSettings.Tests/Autosave/CollectionReloadAndMixinBindingTests.cs
@@ -5,8 +5,6 @@ using System.Runtime.CompilerServices;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nucs.JsonSettings.Autosave;
-using Nucs.JsonSettings.Examples;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation.Recovery;
 using Nucs.JsonSettings.NotifyChanges;
 using Nucs.JsonSettings.Tests.Utils;
