+++ tests/JsonSettings.Tests/Autosave/NotifyChangesEdgeTests.cs
@@ -6,7 +6,6 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
 using Nucs.JsonSettings.Autosave;
 using Nucs.JsonSettings.NotifyChanges;
-using Nucs.JsonSettings.Examples;
 
 namespace Nucs.JsonSettings.Tests.Autosave {
     /// <summary>
