+++ tests/JsonSettings.Tests/Autosave/NotifyChangesMarshalingTests.cs
@@ -5,7 +5,6 @@ using System.Threading;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nucs.JsonSettings.NotifyChanges;
-using Nucs.JsonSettings.Examples;
 using Nucs.JsonSettings.Tests.Utils;
 
 namespace Nucs.JsonSettings.Tests.Autosave {
