+++ tests/JsonSettings.Tests/Autosave/NotifyChangingTests.cs
@@ -3,7 +3,6 @@ using System.ComponentModel;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nucs.JsonSettings.NotifyChanges;
-using Nucs.JsonSettings.Examples;
 using Nucs.JsonSettings.Tests.Utils;
 
 namespace Nucs.JsonSettings.Tests.Autosave {
