+++ tests/JsonSettings.Tests/ConfigurableTests.cs
@@ -1,6 +1,5 @@
 ﻿using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation;
 using Nucs.JsonSettings.Tests.Utils;
