+++ tests/JsonSettings.Tests/FluentApiCoverageTests.cs
@@ -4,7 +4,6 @@ using System.Security;
 using System.Text;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation;
 using Nucs.JsonSettings.Tests.Utils;
