+++ tests/JsonSettings.Tests/HowToUse.cs
@@ -1,6 +1,5 @@
 ﻿using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Tests.Utils;
