+++ tests/JsonSettings.Tests/LegacyEncryptionParityTests.cs
@@ -4,7 +4,6 @@ using System.Security.Cryptography;
 using System.Text;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation;
 using Nucs.JsonSettings.Modulation.Encryption;
 using Nucs.JsonSettings.Tests.Utils;
