+++ tests/JsonSettings.Tests/ObsoleteRijndaelModuleTests.cs
@@ -3,7 +3,6 @@ using System.Security;
 using System.Text;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation;
 using Nucs.JsonSettings.Tests.Utils;
