+++ tests/JsonSettings.Tests/Upgrade/DamagedEncryptedFileTests.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Text;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Modulation.Recovery;
 using Nucs.JsonSettings.Tests.Utils;
