+++ tests/JsonSettings.Tests/Upgrade/ModuleChainingTests.cs
@@ -2,7 +2,6 @@ using System;
 using System.IO;
 using AwesomeAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Nucs.JsonSettings.Fluent;
 using Nucs.JsonSettings.Tests.Utils;
 
 namespace Nucs.JsonSettings.Tests.Upgrade {
